Move QuickHull algo in its own source file
