+++ convex-hull/include/QuickHull.h
@@ -1,0 +1,7 @@
+﻿#ifndef QUICKHULL_H
+#define QUICKHULL_H
+#include "VisualAlgorithm.h"
+
+AlgorithmGenerator quickHullAlgorithm(const INPUT_PARAMETER& points);
+
+#endif //QUICKHULL_H
